fix bugs

+++ src/main/java/org/lightjason/agentspeak/language/execution/expression/numerical/CMultiplicative.java
@@ -120,8 +120,8 @@ public final class CMultiplicative extends IBaseBinary
     private <N extends Number, M extends Number> Number multiply( final N p_left, final M p_right )
     {
         return ( p_left instanceof Double ) || ( p_right instanceof Double )
-               ? p_left.doubleValue() * p_right.doubleValue()
-               : p_left.longValue() * p_right.longValue();
+               ? Double.valueOf( p_left.doubleValue() * p_right.doubleValue() )
+               : Long.valueOf( p_left.longValue() * p_right.longValue() );
     }
 
     /**
@@ -137,8 +137,8 @@ public final class CMultiplicative extends IBaseBinary
     private <N extends Number, M extends Number> Number divide( final N p_left, final M p_right )
     {
         return ( p_left instanceof Double ) || ( p_right instanceof Double )
-               ? p_left.doubleValue() / p_right.doubleValue()
-               : p_left.longValue() / p_right.longValue();
+               ? Double.valueOf( p_left.doubleValue() / p_right.doubleValue() )
+               : Long.valueOf( p_left.longValue() / p_right.longValue() );
     }
 
     /**
